Added new module to control the Readspeak.
Module consists of some overrides and a template file to display Readspeak options.

TODO: add a new variable in the publishing options section of every content type to determine
if Readspeaker is on or off - this is to save JS load on every page. Will be on by default.
Styling of the Readspeaker button will also need some work.