Added options to display more link on testimonial
Removed the helper function and merged into the module for brevity.
Used the field theme function to set variables on the node that could be used
by the node template to control the button display.
I cried a bit too as this hurt my head but was very simple really.